Move the base test to .common
